moved extraction of data from proof to separate module
