+++ Cargo.toml
@@ -2,43 +2,43 @@
 resolver = "2"
 
 members = [
-  "bin/dojo-language-server",
-  "bin/katana",
-  "bin/saya",
-  "bin/scheduler",
-  "bin/sozo",
-  "bin/torii",
-  "crates/benches",
-  "crates/common",
-  "crates/dojo-bindgen",
-  "crates/dojo-core",
-  "crates/dojo-lang",
-  "crates/dojo-test-utils",
-  "crates/dojo-types",
-  "crates/dojo-world",
-  "crates/dojo-world/abigen",
-  "crates/katana/core",
-  "crates/katana/executor",
-  "crates/katana/primitives",
-  "crates/katana/rpc/rpc",
-  "crates/katana/rpc/rpc-api",
-  "crates/katana/rpc/rpc-types",
-  "crates/katana/rpc/rpc-types-builder",
-  "crates/katana/runner",
-  "crates/katana/runner/runner-macro",
-  "crates/katana/storage/codecs",
-  "crates/katana/storage/codecs/derive",
-  "crates/katana/storage/db",
-  "crates/katana/storage/provider",
-  "crates/katana/tasks",
-  "crates/metrics",
-  "crates/saya/core",
-  "crates/saya/provider",
-  "crates/sozo/signers",
-  "crates/torii/client",
-  "crates/torii/server",
-  "crates/torii/types-test",
-  "examples/spawn-and-move",
+    "bin/dojo-language-server",
+    "bin/katana",
+    "bin/saya",
+    "bin/scheduler",
+    "bin/sozo",
+    "bin/torii",
+    "crates/benches",
+    "crates/common",
+    "crates/dojo-bindgen",
+    "crates/dojo-core",
+    "crates/dojo-lang",
+    "crates/dojo-test-utils",
+    "crates/dojo-types",
+    "crates/dojo-world",
+    "crates/dojo-world/abigen",
+    "crates/katana/core",
+    "crates/katana/executor",
+    "crates/katana/primitives",
+    "crates/katana/rpc/rpc",
+    "crates/katana/rpc/rpc-api",
+    "crates/katana/rpc/rpc-types",
+    "crates/katana/rpc/rpc-types-builder",
+    "crates/katana/runner",
+    "crates/katana/runner/runner-macro",
+    "crates/katana/storage/codecs",
+    "crates/katana/storage/codecs/derive",
+    "crates/katana/storage/db",
+    "crates/katana/storage/provider",
+    "crates/katana/tasks",
+    "crates/metrics",
+    "crates/saya/core",
+    "crates/saya/provider",
+    "crates/sozo/signers",
+    "crates/torii/client",
+    "crates/torii/server",
+    "crates/torii/types-test",
+    "examples/spawn-and-move",
 ]
 
 [workspace.package]
